+++ chainstate/src/detail/orphan_blocks/mod.rs
@@ -22,8 +22,8 @@ pub use pool::*;
 mod orphans_proxy;
 pub use orphans_proxy::*;
 
-mod orphans_proxy_impls;
-pub use orphans_proxy_impls::*;
+mod orphans_proxy_impl;
+pub use orphans_proxy_impl::*;
 
-mod pool_impls;
-pub use pool_impls::*;
+mod pool_impl;
+pub use pool_impl::*;
